fix transaksi tambah

+++ page/data_transaksi/transaksi/display.php
@@ -30,24 +30,26 @@
                             </thead>
                             <tbody>
                                 <?php $data = query("SELECT * FROM tb_transaksi");
-                                $nomor = 0;
-                                foreach ($data as $d) : ?>
-                                    <tr>
-                                        <td style="white-space: nowrap; text-align:center;"><?= ++$nomor; ?></td>
-                                        <td style="white-space: nowrap; text-align:left;"><?= $d['konsumen'];; ?></td>
-                                        <td style="white-space: nowrap; text-align:left;"><?= $d['kode_transaksi']; ?></td>
-                                        <td style="white-space: nowrap; text-align:left;"><?= $d['kode_barang']; ?></td>
-                                        <td style="white-space: nowrap; text-align:left;"><?= $d['barang']; ?></td>
-                                        <td style="white-space: nowrap; text-align:right;"><?= $d['Qty']; ?></td>
-                                        <td style="white-space: nowrap; text-align:right;">Rp. <?= number_format($d['Harga'], 0, ',', '.'); ?></td>
-                                        <td style="white-space: nowrap; text-align:right;">Rp. <?= number_format($d['total_harga'], 0, ',', '.'); ?></td>
-                                        <td style="white-space: nowrap; text-align:left;"><?= $d['tgl']; ?></td>
-                                        <td style="white-space: nowrap; text-align:center;">
-                                            <a href="<?= $_baseurl; ?>&aksi=ubah&id=<?= $d['no']; ?>" class="btn btn-info"><i class="fa fa-edit"></i> ubah</a>
-                                            <a onclick="return confirm('Anda yakin ingin menghapus?')" href="<?= $_baseurl; ?>&aksi=hapus&id=<?php echo $d['no']; ?>" class="btn btn-danger"><i class="fa fa-trash"></i> hapus</a>
-                                        </td>
-                                    </tr>
-                                <?php endforeach; ?>
+                                if ($data) :
+                                    $nomor = 0;
+                                    foreach ($data as $d) : ?>
+                                        <tr>
+                                            <td style="white-space: nowrap; text-align:center;"><?= ++$nomor; ?></td>
+                                            <td style="white-space: nowrap; text-align:left;"><?= $d['konsumen'];; ?></td>
+                                            <td style="white-space: nowrap; text-align:left;"><?= $d['kode_transaksi']; ?></td>
+                                            <td style="white-space: nowrap; text-align:left;"><?= $d['kode_barang']; ?></td>
+                                            <td style="white-space: nowrap; text-align:left;"><?= $d['barang']; ?></td>
+                                            <td style="white-space: nowrap; text-align:right;"><?= $d['Qty']; ?></td>
+                                            <td style="white-space: nowrap; text-align:right;">Rp. <?= number_format($d['Harga'], 0, ',', '.'); ?></td>
+                                            <td style="white-space: nowrap; text-align:right;">Rp. <?= number_format($d['total_harga'], 0, ',', '.'); ?></td>
+                                            <td style="white-space: nowrap; text-align:left;"><?= $d['tgl']; ?></td>
+                                            <td style="white-space: nowrap; text-align:center;">
+                                                <a href="<?= $_baseurl; ?>&aksi=ubah&id=<?= $d['no']; ?>" class="btn btn-info"><i class="fa fa-edit"></i> ubah</a>
+                                                <a onclick="return confirm('Anda yakin ingin menghapus?')" href="<?= $_baseurl; ?>&aksi=hapus&id=<?php echo $d['no']; ?>" class="btn btn-danger"><i class="fa fa-trash"></i> hapus</a>
+                                            </td>
+                                        </tr>
+                                <?php endforeach;
+                                endif; ?>
                             </tbody>
                         </table>
                     </div>
